[Refactor] 패키지 깔끔하게 정리
