+++ src/main/java/com/bbangle/bbangle/member/domain/Member.java
@@ -3,7 +3,7 @@ package com.bbangle.bbangle.member.domain;
 import com.bbangle.bbangle.common.domain.BaseEntity;
 import com.bbangle.bbangle.member.customer.dto.InfoUpdateRequest;
 import com.bbangle.bbangle.member.customer.dto.MemberInfoRequest;
-import com.bbangle.bbangle.member.exception.UserValidator;
+import com.bbangle.bbangle.member.customer.exception.UserValidator;
 import com.bbangle.bbangle.token.oauth.domain.OauthServerType;
 import com.bbangle.bbangle.wishlist.domain.WishListFolder;
 import com.bbangle.bbangle.wishlist.domain.WishListStore;
