+++ src/main/java/com/bbangle/bbangle/review/repository/ReviewRepositoryImpl.java
@@ -14,8 +14,8 @@ import com.bbangle.bbangle.review.customer.dto.QLikeCountPerReviewIdDto;
 import com.bbangle.bbangle.review.customer.dto.QReviewCountPerBoardIdDto;
 import com.bbangle.bbangle.review.customer.dto.QReviewDto;
 import com.bbangle.bbangle.review.customer.dto.QReviewSingleDto;
-import com.bbangle.bbangle.review.dao.QReviewStatisticDao;
-import com.bbangle.bbangle.review.dao.ReviewStatisticDao;
+import com.bbangle.bbangle.review.repository.dao.QReviewStatisticDao;
+import com.bbangle.bbangle.review.repository.dao.ReviewStatisticDao;
 import com.bbangle.bbangle.review.domain.QReview;
 import com.bbangle.bbangle.review.domain.QReviewLike;
 import com.bbangle.bbangle.review.domain.ReviewCursor;
