+++ src/main/java/com/bbangle/bbangle/survey/customer/collections/DietLimitations.java
@@ -1,7 +1,7 @@
-package com.bbangle.bbangle.survey.collections;
+package com.bbangle.bbangle.survey.customer.collections;
 
 import com.bbangle.bbangle.survey.domain.DietLimitationInfo;
-import com.bbangle.bbangle.survey.enums.DietLimitation;
+import com.bbangle.bbangle.survey.domain.enums.DietLimitation;
 import java.util.ArrayList;
 import java.util.List;
