+++ src/main/java/com/bbangle/bbangle/survey/customer/controller/SurveyController.java
@@ -3,11 +3,11 @@ package com.bbangle.bbangle.survey.customer.controller;
 import com.bbangle.bbangle.common.dto.CommonResult;
 import com.bbangle.bbangle.common.dto.SingleResult;
 import com.bbangle.bbangle.common.service.ResponseService;
-import com.bbangle.bbangle.survey.collections.DietLimitations;
-import com.bbangle.bbangle.survey.collections.HealthConcerns;
-import com.bbangle.bbangle.survey.collections.IsVegetarians;
-import com.bbangle.bbangle.survey.collections.SurveyInfo;
-import com.bbangle.bbangle.survey.collections.UnmatchedIngredients;
+import com.bbangle.bbangle.survey.customer.collections.DietLimitations;
+import com.bbangle.bbangle.survey.customer.collections.HealthConcerns;
+import com.bbangle.bbangle.survey.customer.collections.IsVegetarians;
+import com.bbangle.bbangle.survey.customer.collections.SurveyInfo;
+import com.bbangle.bbangle.survey.customer.collections.UnmatchedIngredients;
 import com.bbangle.bbangle.survey.customer.dto.request.RecommendationSurveyDto;
 import com.bbangle.bbangle.survey.customer.service.SurveyService;
 import lombok.RequiredArgsConstructor;
