+++ src/main/java/com/bbangle/bbangle/survey/customer/service/SurveyService.java
@@ -2,7 +2,7 @@ package com.bbangle.bbangle.survey.customer.service;
 
 import com.bbangle.bbangle.exception.BbangleErrorCode;
 import com.bbangle.bbangle.exception.BbangleException;
-import com.bbangle.bbangle.survey.collections.SurveyInfo;
+import com.bbangle.bbangle.survey.customer.collections.SurveyInfo;
 import com.bbangle.bbangle.survey.domain.DietLimitationInfo;
 import com.bbangle.bbangle.survey.domain.FoodSurvey;
 import com.bbangle.bbangle.survey.domain.HealthConcernInfo;
