+++ src/main/java/com/bbangle/bbangle/survey/domain/enums/IsVegetarian.java
@@ -1,4 +1,4 @@
-package com.bbangle.bbangle.survey.enums;
+package com.bbangle.bbangle.survey.domain.enums;
 
 import lombok.RequiredArgsConstructor;
